Fix technology constants in switch in iframe.php
In iframe.php, bare text 'case HTML5' is used inside a switch
statement. This is incorrect as there is no 'HTML5' constant
defined in the code. It needs quotes around the 'HTML5'.

It works correctly already because PHP defaults to treating an
unrecognised constant as the same as the string. However, this
causes warning messages in the PHP log.